Catch SQL exceptions properly.
If we allow an SPI function to throw an SQL exception, the Lua stack is
not properly unwound and we get a mess. Put a Pg TRY/CATCH block around
every SPI function that might throw, so we can convert the SQL exception
into a Lua exception and rethrow it.

Currently the Lua exception is just a string. It would be better to turn
the ErrorData into a Lua userdata with proper methods, so if it runs off
the top of the Lua stack it can be rethrown in its original form.