Change the format of Lua->SQL error conversion.
Put the real error message in the message part, and the 'this is a pllua
error' part in the detail. Otherwise the real error message tends to get
lost, as client libraries other PLs don't provide access to the detail
part of the error.

Also throw Lua-originated errors with a new made-up SQLSTATE of 'P0LUA'.
'P0' is the prefix for PL/pgSQL errors, so these should be handled
reasonably by code ignorant of PL/Lua, but it seems better than the
DATA_EXCEPTION code used previously.