Throw SQL ERRORs as sqlerr userdata.
This means Lua code can get at all the fields, and, if the error is
eventually rethrown, it will include all the original information.

The sqlerr userdata is subject to Lua's gc, so it's important to copy
the ErrorData it points to into PL/Lua's memory context. Otherwise it
will end up in the context for this function call, which will probably
be destroyed before Lua gets round to gcing the userdata.