Export an 'isudata' function to Lua.
This allows Lua code to identify the type of a PL/Lua userdata;
surprisingly, Lua doesn't provide any native mechanism for this. This
means maintaining a registry of the names of the PL/Lua userdatas.

The reason for making this a predicate rather than a function returning
a type name is that it allows for the possibility of subclassing later
on.